+++ MyWeather/MyWeather/View Controllers/ViewController.swift
@@ -5,9 +5,8 @@
 //  Created by Danil Nurgaliev on 20.05.2021.
 //
 
-import UIKit
-
 import CoreLocation
+import UIKit
 
 class ViewController: UIViewController {
     private var latitude: CLLocationDegrees?
@@ -49,10 +48,36 @@ class ViewController: UIViewController {
 
     @IBAction private func locationButtonPressed(_ sender: UIButton) {
         guard let latitude = latitude, let longitude = longitude else {
+            assertionFailure("no latitude no longitude")
             return
         }
 
         networkWeatherManager.getCurrentWeather(forRequestType: .coordinate(latitude: latitude, longitude: longitude))
+    }
+
+    private func presentSearchAlertController(title: String?, message: String?, completionHandler: @escaping (String) -> Void) {
+        let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        alertController.addTextField { textField in
+            let cities = ["Kazan", "Minsk", "Moscow", "Stambul", "New York"]
+            textField.placeholder = cities.randomElement()
+        }
+
+        let search = UIAlertAction(title: "Search", style: .default) { _ in
+            let textfield = alertController.textFields?.first
+            guard  let cityName = textfield?.text else {
+                return
+            }
+
+            if !cityName.isEmpty {
+                let city = cityName.split(separator: " ").joined(separator: "%20")
+                completionHandler(city)
+            }
+        }
+        let cancel = UIAlertAction(title: "Cancel", style: .default, handler: nil)
+
+        alertController.addAction(cancel)
+        alertController.addAction(search)
+        present(alertController, animated: true, completion: nil)
     }
 }
 
@@ -75,7 +100,11 @@ extension ViewController: CLLocationManagerDelegate {
     func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
         let location = locations.last
 
-        guard let latitude = location?.coordinate.latitude, let longitude = location?.coordinate.longitude else {
+        latitude = location?.coordinate.latitude
+        longitude = location?.coordinate.longitude
+
+        guard let latitude = latitude, let longitude = longitude else {
+            assertionFailure("no latitude no longitude")
             return
         }
